move utility functions from Batfish to CommonUtil
