+++ projects/batfish/src/main/java/org/batfish/grammar/routing_table/ios/IosRoutingTableExtractor.java
@@ -2,7 +2,6 @@ package org.batfish.grammar.routing_table.ios;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
@@ -12,6 +11,7 @@ import org.antlr.v4.runtime.tree.ParseTreeWalker;
 import org.batfish.common.BatfishException;
 import org.batfish.common.Warnings;
 import org.batfish.common.plugin.IBatfish;
+import org.batfish.common.util.CommonUtil;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.Interface;
@@ -63,8 +63,7 @@ public class IosRoutingTableExtractor extends IosRoutingTableParserBaseListener
     _parser = parser;
     _w = w;
     Map<String, Configuration> configurations = batfish.loadConfigurations();
-    Map<Ip, Set<String>> ipOwners = batfish.computeIpOwners(configurations, true);
-    Map<Ip, String> ipOwnersSimple = batfish.computeIpOwnersSimple(ipOwners);
+    Map<Ip, String> ipOwnersSimple = CommonUtil.computeIpOwnersSimple(configurations, true);
     _ipOwners = ipOwnersSimple;
   }
